requirements.txt
gradio==5.40.0
pydantic==2.11.7
fastapi==0.116.1
openai==1.98.0
tiktoken==0.9.0
langchain==0.3.27
langchain-openai==0.3.28
google-search-results==2.4.2
python-docx==1.2.0
openpyxl==3.1.5
python-dotenv==1.1.1
uvicorn==0.35.0
tabulate==0.9.0
pandas==2.3.1
numpy==2.3.2
requests==2.32.4
